Correciones realizadas
