+++ next.config.js
@@ -7,6 +7,15 @@ const nextConfig = {
   typescript: {
     ignoreBuildErrors: true,
   },
+  async redirects() {
+    return [
+      {
+        source: "/",
+        destination: "/ventas",
+        permanent: false,
+      },
+    ];
+  },
 };
 
 module.exports = nextConfig
