--- src/api/brands.server.js
@@ -1,7 +1,0 @@
-import { getServerEntityById } from "./server";
-
-export const getBrandById = (id) => getServerEntityById({
-  id,
-  path: "brands",
-  responseEntity: "brand",
-});
