--- src/api/products.server.js
@@ -1,7 +1,0 @@
-import { getServerEntityById } from "./server";
-
-export const getProductById = (id) => getServerEntityById({
-  id,
-  path: "products",
-  responseEntity: "product",
-});
