--- src/api/suppliers.server.js
@@ -1,7 +1,0 @@
-import { getServerEntityById } from "./server";
-
-export const getSupplierById = (id) => getServerEntityById({
-  id,
-  path: "suppliers",
-  responseEntity: "supplier",
-});
