+++ src/app/productos/[id]/page.jsx
@@ -1,20 +1,10 @@
-import { getProductById } from "@/api/products.server";
-import { PAGE_CONSTANTS } from "@/common/constants/pages";
-import { redirect } from "next/navigation";
-import ProductPage from "./ProductPage";
+import { getEntityById } from "@/api/entity";
+import PageClient from "./page.client";
 
 const Product = async ({ params }) => {
-  const { product, status } = await getProductById(params?.id);
+  const product = await getEntityById({ id: params?.id, path: "products", responseEntity: "product" });
 
-  if ([401, 403].includes(status)) {
-    redirect(PAGE_CONSTANTS.LOGIN.BASE);
-  }
-
-  if (!product) {
-    redirect(PAGE_CONSTANTS.NOT_FOUND.BASE);
-  }
-
-  return <ProductPage product={product} />;
+  return <PageClient product={product} />;
 };
 
 export default Product;
