+++ src/app/ventas/[id]/page.jsx
@@ -1,24 +1,21 @@
-import { getBudgetById } from "@/api/budgets.server";
+import { getEntityById } from "@/api/entity";
 import { PAGE_CONSTANTS } from "@/common/constants/pages";
 import { redirect } from "next/navigation";
-import BudgetPage from "./BudgetPage";
+import PageClient from "./page.client";
 
 const Budget = async ({ params }) => {
-  const { budget, status } = await getBudgetById(params?.id);
-
-  if ([401, 403].includes(status)) {
-    redirect(PAGE_CONSTANTS.LOGIN.BASE);
-  }
-
-  if (!budget) {
-    redirect(PAGE_CONSTANTS.NOT_FOUND.BASE);
-  }
+  const budgetResponse = await getEntityById({ id: params?.id, path: "budgets", responseEntity: "budget" });
+  const budget = {
+    ...budgetResponse,
+    globalDiscount: budgetResponse.globalDiscount ?? 0,
+    additionalCharge: budgetResponse.additionalCharge ?? 0,
+  };
 
   if (budget.state === "DRAFT") {
     redirect(`${PAGE_CONSTANTS.BUDGETS.BASE}/${params?.id}/${PAGE_CONSTANTS.BUDGETS.DRAFT_SEGMENT}`);
   }
 
-  return <BudgetPage budget={budget} />;
+  return <PageClient budget={budget} />;
 };
 
 export default Budget;
